% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateJunctionReadCounts.R
\name{calculateJunctionReadCounts}
\alias{calculateJunctionReadCounts}
\title{Calculate sum of 1st exon-intron junction reads overlapping input promoter loci}
\usage{
calculateJunctionReadCounts(junctionFile = "", junctionType,
  promoterFile)
}
\arguments{
\item{junctionFile}{Path to junction files either from STAR or tophat}

\item{junctionType}{character type of the junction bed file. Either 'tophat' or 'star'}

\item{promoterFile}{full path file name of the promoter loci - should be a bed file. chr start stop as the 1st 3 columns in zero based format}
}
\value{
A data.frame object with 1st exon junction reads for each promoter loci
}
\description{
Calculate sum of 1st exon-intron junction reads overlapping input promoter loci
}
\examples{
\dontrun{
junction_reads <- calculateJunctionReadCounts(junctionFile="test/abc.junction.bed",
                                   junctionType="tophat",
                                   promoterFile="test/aploci.bed")




}


}
\author{
Aditi Qamra
}
