% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-survival.R
\name{calculateSurvival}
\alias{calculateSurvival}
\title{Given alternate promoter score for samples,
divide samples into groups and calculate survival differences}
\usage{
calculateSurvival(alternatePromoterScore, clinicaldata,
  alternatePromoterQuantileThreshold, survivalTime, survivalEvent, title,
  ...)
}
\arguments{
\item{alternatePromoterScore}{A data.frame object. Two columns containing sample ids and
numeric alternate promoter score}

\item{clinicaldata}{A data.frame object. Should have an id column labelled "id"
and have survival time and event information}

\item{survivalTime}{Column in clinicaldata specifying survival time}

\item{survivalEvent}{Column in clinicaldata specifying event information}

\item{title}{Character type. Label to plot on survival curve}

\item{alternatePromoterUpperQuantileThreshold}{Numeric value greater than 0 and less than 1
Identifying threshold to define samples into APhigh and APlow.
APhigh samples will be alternatePromoterScore >= quantile(alternatePromoterScore,alternatePromoterUpperQuantileThreshold)
APlow samples will be alternatePromoterScore <= quantile(alternatePromoterScore,(1-alternatePromoterUpperQuantileThreshold))}
}
\value{
A data.frame object.Return @param alternatePromoterScore with AP group classification
}
\description{
Given alternate promoter score for samples,
divide samples into groups and calculate survival differences
}
\examples{
\dontrun{
alternatePromoterGroups <- calculateSurvival(alternatePromoterScore,
                                                         clinicaldata,
                                                         alternatePromoterUpperQuantileThreshold=0.66,
                                                         survivalTime="pfs",
                                                         survivalEvent="progression",
                                                         title="AP Group -  66/33 Split")
}

}
\seealso{
\code{\link{calculateAlternatePromoterScore}} for calculating
alternate promoter score for samples
}
