% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-coverage-matrix.R
\name{createCoverageMatrix}
\alias{createCoverageMatrix}
\title{Create normalized data frame from  input coverage bed files 
The function will read in bed files, coverage column will be specified by user
and collate them to make a data matrix and then normalize by the library size. 
If the library sizes are not provided by then use sum of reads across loci as proxy. 
Though this is not recommended.}
\usage{
createCoverageMatrix(inputPath, filePattern, junctionReads, junctionType,
  promoterFile, normalize, librarySizeFile, ...)
}
\arguments{
\item{inputPath}{Path to coverage or junction bed files. 
Assumptions for coverage bed file - 1st 3 columns are the loci of interest (chr start stop) and the last column is the coverage readcount}

\item{filePattern}{character string or regular expression to search inputPath to get all coverage files}

\item{junctionReads}{logical flag specifying whether junction reads are to be calculated OR not}

\item{junctionType}{character type of the junction bed file. Either 'tophat' or 'star'}

\item{promoterFile}{Full path file name of promoter loci - needs to be a bed file if junctionReads flag is equal to TRUE}

\item{normalize}{logical flag specifying whether normalization by library size should be done or not}

\item{librarySizeFile}{full path file name with library sizes. Should have two columns
1st should be sample name - same as file name in @param inputPath and the 2nd as the library size. No of rows should be equal to
no of coverage files read from @param inputpath}
}
\value{
A data.frame object with normalized RNAseq reads in each sample (rows)
}
\description{
Create normalized data frame from  input coverage bed files 
The function will read in bed files, coverage column will be specified by user
and collate them to make a data matrix and then normalize by the library size. 
If the library sizes are not provided by then use sum of reads across loci as proxy. 
Though this is not recommended.
}
\examples{
\dontrun{
norm_matrix <- createCoverageMatrix(inputPath="readcount/",
                                   filePattern="coverage_",
                                   junctionReads=FALSE,
                                   junctionType='',
                                   promoterFile='',
                                   normalize=TRUE,
                                   librarySizeFile="libsizes.txt")

norm_matrix <- createCoverageMatrix(inputPath="test/",
                                   filePattern=".junction.bed",
                                   junctionReads=TRUE,
                                   junctionType='tophat',
                                   promoterFile='test/aploci.bed',
                                   normalize=TRUE,
                                   librarySizeFile="libsizes.txt")



}


}
\author{
Aditi Qamra
}
