% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gencode.v19.intron.first.dedup}
\alias{gencode.v19.intron.first.dedup}
\title{Ist intron dataset for Gencode.v19}
\format{An object of class \code{GRanges} of length 117581.}
\usage{
gencode.v19.intron.first.dedup
}
\description{
A GRanges object containing the coordinates for 1st exon for gencode v19
}
\keyword{datasets}
