% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bed.R
\name{read_bed}
\alias{read_bed}
\title{ReadBed function from package genomation - https://github.com/BIMSBbioinfo/genomation/blob/master/R/readData.R}
\usage{
read_bed(file, track.line = FALSE, remove.unusual = FALSE,
  zero.based = TRUE)
}
\value{
a granges object
}
\description{
ReadBed function from package genomation - https://github.com/BIMSBbioinfo/genomation/blob/master/R/readData.R
}
