% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-alternatepromoter-score.R
\name{calculateAlternatePromoterScore}
\alias{calculateAlternatePromoterScore}
\title{Calculate AP score given a readcount matrix and H3k4me3 annotation}
\usage{
calculateAlternatePromoterScore(promoterReadCounts, promoterAnnotation,
  promoterMethod = c("medianbased", "rankbased"), medianThreshold, ...)
}
\arguments{
\item{promoterReadCounts}{A data.frame object. The number of reads
per promoter (rows) for each sample (cols). Rownames should be promoter names}

\item{promoterAnnotation}{A list with elements. Number of elements
equal to nrow @param (promoterReadCounts) and order same as @param promoterReadCounts
indicating directionality i.e. 'gain' or 'loss'
of promoters in @param promoterReadCounts}

\item{promoterMethod}{Character type.
Identifying method of AP score calculation
Can be either "medianbased" or "rankbased"}

\item{medianThreshold}{A numeric value.
Specify if @param promoterMethod is equal to "medianbased"}
}
\value{
A data.frame object. An AP score (column) calculated for each sample (rows) 
using method @param promoterMethod
}
\description{
Calculate AP score given a readcount matrix and H3k4me3 annotation
}
\examples{
\dontrun{
alternatePromoterScore <- calculateAlternatePromoterScore(promoterReadCounts,
                                                         promoterAnnotation,
                                                         promoterMethod = "medianbased",
                                                         medianThreshold = 4)
}

}
\author{
Aditi Qamra
}
